+++ tests/Unit/Messaging/CloudMessageTest.php
@@ -6,6 +6,8 @@ namespace Kreait\Firebase\Tests\Unit\Messaging;
 
 use Kreait\Firebase\Exception\InvalidArgumentException;
 use Kreait\Firebase\Messaging\CloudMessage;
+use Kreait\Firebase\Messaging\MessageTarget;
+use Kreait\Firebase\Messaging\Notification;
 use PHPUnit\Framework\TestCase;
 
 class CloudMessageTest extends TestCase
@@ -13,6 +15,22 @@ class CloudMessageTest extends TestCase
     public function testInvalidTargetCausesError()
     {
         $this->expectException(InvalidArgumentException::class);
-        CloudMessage::withTarget('foo', 'bar');
+        CloudMessage::withTarget(MessageTarget::TOKEN, 'bar');
+    }
+
+    public function testWithChangedTarget()
+    {
+        $original = CloudMessage::withTarget(MessageTarget::TOKEN, 'bar')
+            ->withData(['foo' => 'bar'])
+            ->withNotification(Notification::create('title', 'body'));
+
+        $changed = $original->withChangedTarget(MessageTarget::TOKEN, 'baz');
+
+        $encodedOriginal = json_decode(json_encode($original), true);
+        $encodedOriginal[MessageTarget::TOKEN] = 'baz';
+
+        $encodedChanged = json_decode(json_encode($changed), true);
+
+        $this->assertSame($encodedOriginal, $encodedChanged);
     }
 }